requirements.txt
Box2D==2.3.10
certifi==2024.8.30
charset-normalizer==2.0.12
cloudpickle==1.6.0
docopt==0.6.2
future==0.16.0
gym == 0.19.0
idna==3.10
numpy==1.13.1
pygame==2.6.1
pyglet==1.5.16
requests==2.27.1
scipy==0.19.1
urllib3==1.26.20
yarg==0.1.10
torch~=1.10.2
mageio==2.15.0
matplotlib==3.7.5
numpy==1.24.4
tensorflow==2.7.0
torch==2.4.1
